+++ packages/fiori-app-sub-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/fiori-app-sub-generator
 
+## 0.11.16
+
+### Patch Changes
+
+- Updated dependencies [5d2a018]
+    - @sap-ux/odata-service-inquirer@2.14.1
+
 ## 0.11.15
 
 ### Patch Changes